Replaced nextFloat() with nextDouble()

+++ src/Example.java
@@ -6,7 +6,7 @@ public class Example {
 
     public static void main(String[] args) throws Exception {
         System.out.print("Please enter a number you want to find the absolute value of: ");
-        double value = scanner.nextFloat();
+        double value = scanner.nextDouble();
         double absoluteValue = findAbsoluteValue.findValue(value);
         System.out.println("|" + value + "| = " + absoluteValue);
         System.out.println("The absolute value of " + value + " is " + absoluteValue);
